Separate Lerp from ColorGradient
