+++ Basestation_Software.Web/Core/Components/GPS.razor
@@ -247,7 +247,6 @@
 
     // Declare page member variables.
     private string jsInteropName;
-    private Random random;
 
     protected override void OnInitialized()
     {
